+++ src/hooks/shortcutsHooks.ts
@@ -1,3 +1,5 @@
+// SHORTCUT HOOKS
+
 import { useDispatch } from "react-redux"
 import { useNavigate } from "react-router-dom"
 import { ROUTES } from "../routes"
